+++ airflow/example_dags/example_dataset_alias.py
@@ -38,7 +38,7 @@ from __future__ import annotations
 import pendulum
 
 from airflow import DAG
-from airflow.datasets import Dataset, DatasetAlias
+from airflow.assets import Dataset, DatasetAlias
 from airflow.decorators import task
 
 with DAG(
