+++ airflow/example_dags/example_dataset_alias_with_no_taskflow.py
@@ -36,7 +36,7 @@ from __future__ import annotations
 import pendulum
 
 from airflow import DAG
-from airflow.datasets import Dataset, DatasetAlias
+from airflow.assets import Dataset, DatasetAlias
 from airflow.operators.python import PythonOperator
 
 with DAG(
