+++ airflow/example_dags/example_datasets.py
@@ -54,7 +54,7 @@ from __future__ import annotations
 
 import pendulum
 
-from airflow.datasets import Dataset
+from airflow.assets import Dataset
 from airflow.models.dag import DAG
 from airflow.operators.bash import BashOperator
 from airflow.timetables.datasets import DatasetOrTimeSchedule
