+++ airflow/example_dags/example_inlet_event_extra.py
@@ -25,7 +25,7 @@ from __future__ import annotations
 
 import datetime
 
-from airflow.datasets import Dataset
+from airflow.assets import Dataset
 from airflow.decorators import task
 from airflow.models.dag import DAG
 from airflow.operators.bash import BashOperator
