+++ airflow/serialization/serialized_objects.py
@@ -34,9 +34,7 @@ from dateutil import relativedelta
 from pendulum.tz.timezone import FixedTimezone, Timezone
 
 from airflow import macros
-from airflow.callbacks.callback_requests import DagCallbackRequest, SlaCallbackRequest, TaskCallbackRequest
-from airflow.compat.functools import cache
-from airflow.datasets import (
+from airflow.assets import (
     BaseDataset,
     Dataset,
     DatasetAlias,
@@ -44,6 +42,8 @@ from airflow.datasets import (
     DatasetAny,
     _DatasetAliasCondition,
 )
+from airflow.callbacks.callback_requests import DagCallbackRequest, SlaCallbackRequest, TaskCallbackRequest
+from airflow.compat.functools import cache
 from airflow.exceptions import AirflowException, SerializationError, TaskDeferred
 from airflow.jobs.job import Job
 from airflow.models import Trigger
