+++ scripts/ci/pre_commit/check_tests_in_right_folders.py
@@ -33,6 +33,7 @@ POSSIBLE_TEST_FOLDERS = [
     "api",
     "api_connexion",
     "api_internal",
+    "assets",
     "auth",
     "callbacks",
     "charts",
@@ -44,7 +45,6 @@ POSSIBLE_TEST_FOLDERS = [
     "dags",
     "dags_corrupted",
     "dags_with_system_exit",
-    "datasets",
     "decorators",
     "executors",
     "hooks",
