+++ scripts/cov/other_coverage.py
@@ -37,7 +37,7 @@ https://app.codecov.io/github/apache/airflow
     "airflow/callbacks",
     "airflow/config_templates",
     "airflow/dag_processing",
-    "airflow/datasets",
+    "airflow/assets",
     "airflow/decorators",
     "airflow/hooks",
     "airflow/io",
@@ -79,7 +79,7 @@ Other tests to potentially run against the source_file packages:
     "tests/cluster_policies",
     "tests/config_templates",
     "tests/dag_processing",
-    "tests/datasets",
+    "tests/assets",
     "tests/decorators",
     "tests/hooks",
     "tests/io",
