+++ tests/api_connexion/schemas/test_dataset_schema.py
@@ -27,7 +27,7 @@ from airflow.api_connexion.schemas.dataset_schema import (
     dataset_event_schema,
     dataset_schema,
 )
-from airflow.datasets import Dataset
+from airflow.assets import Dataset
 from airflow.models.dataset import DatasetAliasModel, DatasetEvent, DatasetModel
 from airflow.operators.empty import EmptyOperator
 from tests.test_utils.db import clear_db_dags, clear_db_datasets
