+++ tests/models/test_taskinstance.py
@@ -39,7 +39,7 @@ import time_machine
 from sqlalchemy import select
 
 from airflow import settings
-from airflow.datasets import DatasetAlias
+from airflow.assets import DatasetAlias
 from airflow.decorators import task, task_group
 from airflow.example_dags.plugins.workday import AfterWorkdayTimetable
 from airflow.exceptions import (
@@ -2404,7 +2404,7 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_dataset_extra(self, dag_maker, session):
-        from airflow.datasets import Dataset
+        from airflow.assets import Dataset
 
         with dag_maker(schedule=None, session=session) as dag:
 
@@ -2446,7 +2446,7 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_dataset_extra_ignore_different(self, dag_maker, session):
-        from airflow.datasets import Dataset
+        from airflow.assets import Dataset
 
         with dag_maker(schedule=None, session=session):
 
@@ -2468,8 +2468,8 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_dataset_extra_yield(self, dag_maker, session):
-        from airflow.datasets import Dataset
-        from airflow.datasets.metadata import Metadata
+        from airflow.assets import Dataset
+        from airflow.assets.metadata import Metadata
 
         with dag_maker(schedule=None, session=session) as dag:
 
@@ -2518,7 +2518,7 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_dataset_alias(self, dag_maker, session):
-        from airflow.datasets import Dataset, DatasetAlias
+        from airflow.assets import Dataset, DatasetAlias
 
         ds_uri = "test_outlet_dataset_alias_test_case_ds"
         dsa_name_1 = "test_outlet_dataset_alias_test_case_dsa_1"
@@ -2567,7 +2567,7 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_multiple_dataset_alias(self, dag_maker, session):
-        from airflow.datasets import Dataset, DatasetAlias
+        from airflow.assets import Dataset, DatasetAlias
 
         ds_uri = "test_outlet_mdsa_ds"
         dsa_name_1 = "test_outlet_mdsa_dsa_1"
@@ -2628,8 +2628,8 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_dataset_alias_through_metadata(self, dag_maker, session):
-        from airflow.datasets import DatasetAlias
-        from airflow.datasets.metadata import Metadata
+        from airflow.assets import DatasetAlias
+        from airflow.assets.metadata import Metadata
 
         ds_uri = "test_outlet_dataset_alias_through_metadata_ds"
         dsa_name = "test_outlet_dataset_alias_through_metadata_dsa"
@@ -2673,7 +2673,7 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_outlet_dataset_alias_dataset_not_exists(self, dag_maker, session):
-        from airflow.datasets import Dataset, DatasetAlias
+        from airflow.assets import Dataset, DatasetAlias
 
         dsa_name = "test_outlet_dataset_alias_dataset_not_exists_dsa"
         ds_uri = "did_not_exists"
@@ -2713,7 +2713,7 @@ class TestTaskInstance:
 
     @pytest.mark.skip_if_database_isolation_mode  # Does not work in db isolation mode
     def test_inlet_dataset_extra(self, dag_maker, session):
-        from airflow.datasets import Dataset
+        from airflow.assets import Dataset
 
         read_task_evaluated = False
 
@@ -2776,7 +2776,7 @@ class TestTaskInstance:
         session.add_all([ds_model, dsa_model])
         session.commit()
 
-        from airflow.datasets import Dataset, DatasetAlias
+        from airflow.assets import Dataset, DatasetAlias
 
         read_task_evaluated = False
 
@@ -2835,7 +2835,7 @@ class TestTaskInstance:
         session.add(dsa_model)
         session.commit()
 
-        from airflow.datasets import DatasetAlias
+        from airflow.assets import DatasetAlias
 
         with dag_maker(schedule=None, session=session):
 
@@ -2866,7 +2866,7 @@ class TestTaskInstance:
         ],
     )
     def test_inlet_dataset_extra_slice(self, dag_maker, session, slicer, expected):
-        from airflow.datasets import Dataset
+        from airflow.assets import Dataset
 
         ds_uri = "test_inlet_dataset_extra_slice"
 
@@ -2929,7 +2929,7 @@ class TestTaskInstance:
         session.add_all([ds_model, dsa_model])
         session.commit()
 
-        from airflow.datasets import Dataset
+        from airflow.assets import Dataset
 
         with dag_maker(dag_id="write", schedule="@daily", params={"i": -1}, session=session):
 
@@ -2974,7 +2974,7 @@ class TestTaskInstance:
         Test that when a task that produces dataset has ran, that changing the consumer
         dag dataset will not cause primary key blank-out
         """
-        from airflow.datasets import Dataset
+        from airflow.assets import Dataset
 
         with dag_maker(schedule=None, serialized=True) as dag1:
