+++ tests/providers/common/io/datasets/test_file.py
@@ -20,7 +20,7 @@ from urllib.parse import urlsplit, urlunsplit
 
 import pytest
 
-from airflow.datasets import Dataset
+from airflow.assets import Dataset
 from airflow.providers.common.compat.openlineage.facet import Dataset as OpenLineageDataset
 from airflow.providers.common.io.datasets.file import (
     convert_dataset_to_openlineage,
