+++ tests/serialization/test_serialized_objects.py
@@ -30,7 +30,7 @@ from dateutil import relativedelta
 from kubernetes.client import models as k8s
 from pendulum.tz.timezone import Timezone
 
-from airflow.datasets import Dataset, DatasetAlias, DatasetAliasEvent
+from airflow.assets import Dataset, DatasetAlias, DatasetAliasEvent
 from airflow.exceptions import (
     AirflowException,
     AirflowFailException,
